requirements.txt
Django
djangorestframework
asgiref
tzdata
mysqlclient
django-cors-headers
clerk-sdk-python
python-jose
cryptography
requests
python-dotenv
Werkzeug
cachelib
